- Login now takes info from form - Prints info to console
