+++ portal/templates/index.html
@@ -3,7 +3,7 @@
 {% block title %}TSCT Portal{% endblock %}
 
 {% block content %}
-  <form method="post">
+  <form action="/login" method="post">
     <label for="email">Email
       <input type="email" name="email" />
     </label>
